[issue #50] Fix bug that causes Rubberband model to become invalid
The RubberBand model of the equalizer panel could become invalid under
certain conditions. A value of -1 could actually occur for a category if
the carryover is negative and the goodIndexes list contains indexes that
point to 0 values. This is the case if the initial delta is smaller than
the number of available indexes (the first and only valueDelta is 0).
The carry over distribution section now handles this case correctly.