[issue #67] CopyOnWriteArrayList for listeners
The SimpleBandModel uses an unprotected ArrayList for storing the
listeners as well as unprotected r/w access to it. The list is replaced
by a CopyOnWriteArrayList to the disadvantage of a synchronisation of
the list access in good faith that there are far more read accesses to
the listener list than write accesses.