issue #97
New ProtocolStepParamter class for wrapping the parameters of the protocolstep.
Interface change in ProtocolStep: new getParameters() method.
ProtocolStepListener implementation for firing an event when the Parameters of a protocolstep
are changed.

new createParameters() method in AbstractProtocolStep for generating the Parameters.
createParameters() is executed by the ProtocolHandler in an own thread (long tasks are no problem).

The ProtocolStepPanelWrapper now visualizes a waiting screen (ugly by now), and shows the
ProtocolStepPanel only when the createParameters() is done.

copyParams now deprecated.